% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{box_cox}
\alias{box_cox}
\title{Compare the Box-Cox Transformation at Different Levels of \eqn{\lambda}}
\usage{
box_cox(X, Y, lambdas, n = NROW(Y))
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}

\item{Y}{An \eqn{n}-length vector or \eqn{n \times 1} matrix representing the response.}

\item{lambdas}{Vector of values of \eqn{\lambda}. The transformation entails raising \eqn{Y} to the power of \eqn{\lambda}, or logging it in the case of \eqn{\lambda = 0}.}

\item{n}{Number of rows in the design matrix.}
}
\value{
Two-column matrix with the vector of \eqn{\lambda} in one column and the estimated \eqn{SSE} in the other.
}
\description{
Compare the Box-Cox Transformation at Different Levels of \eqn{\lambda}
}
\details{
This function assess the Box-Cox transformation at different values of
\eqn{\lambda}, the tuning parameter. The best corresponds to the lowest model \eqn{SSE}.
}
