% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{breusch_pagan}
\alias{breusch_pagan}
\title{Conduct the Breusch-Pagan Test of Heteroscedascity}
\usage{
breusch_pagan(
  X,
  resid,
  SSE,
  n = NROW(X),
  q = seq_len(NCOL(X)),
  alpha = 0.05,
  has_intercept = TRUE
)
}
\arguments{
\item{X}{Design matrix of a linear model.}

\item{resid}{Vector of residuals from the fitted model.}

\item{SSE}{Sum of squared errors of the fitted model.}

\item{n}{Number of rows of the design matrix. Computed if not provided.}

\item{q}{Number of parameters in the model.}

\item{alpha}{Alpha level of the significance test.}

\item{has_intercept}{Logical. Was the model fitted with an intercept?}
}
\value{
Value of the Breusch-Pagan statistic.
}
\description{
Conduct the Breusch-Pagan Test of Heteroscedascity
}
\details{
This function carries out the Breusch-Pagan test, which conducts a chi-squared test to assess whether the residuals have constant variance.
}
