% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{brown_forsythe}
\alias{brown_forsythe}
\title{Conduct the Brown-Forsythe Test of Equal Variances on Residuals}
\usage{
brown_forsythe(
  resid,
  groups = resid <= stats::median(resid),
  n = NROW(resid),
  p,
  alpha = 0.05
)
}
\arguments{
\item{resid}{vector of model residuals.}

\item{groups}{Factor or logical of the same length as \code{resid} denoting groups into which to split the residuals. Defaults to a logical vector indicating whether each residual is below the median residual.}

\item{n}{Number of rows in the design matrix.}

\item{p}{Number of columns in the design matrix..}

\item{alpha}{Alpha level for the significance test.}
}
\value{
Value of the Brown-Forsythe statistic.
}
\description{
Conduct the Brown-Forsythe Test of Equal Variances on Residuals
}
\details{
This function performs the Brown-Forsythe test of equal variance on specified groups of model residuals. A significant result provides evidence of unequal variances.
}
