% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{calibrate}
\alias{calibrate}
\title{Estimate Observations of \eqn{X} from a Given \code{Y}}
\usage{
calibrate(
  Y,
  Y_0,
  betas,
  Y_bar = mean(Y),
  X,
  m = 1,
  n = NROW(Y),
  p = NROW(betas),
  MSE,
  alpha = 0.05,
  simultaneous = FALSE,
  intercept_fitted = TRUE
)
}
\arguments{
\item{Y}{Vector of \eqn{Y} observations used to fit the model}

\item{Y_0}{Hypothetical \eqn{Y} observations on which to predict \code{X}}

\item{betas}{Estimated regression coefficients.}

\item{Y_bar}{Mean of the response vector \eqn{Y}. Will be computed if not provided.}

\item{X}{Design matrix used to fit a model.}

\item{m}{Number of new observations at each level of \code{Y_0} to predict.}

\item{n}{Number of rows in the design matrix.}

\item{p}{Number of parameters in the model.}

\item{MSE}{Estimated mean squared error.}

\item{alpha}{Alpha level for confidence intervals.}

\item{simultaneous}{Logical. Should Bonferroni-corrected simultaneous intervals be constructed for each prediction? Default \code{FALSE}.}

\item{intercept_fitted}{Logical. Does the model contain an intercept parameter? Default \code{FALSE}.}
}
\value{
\eqn{3 \times h} matrix of predictions with interval bounds, where \eqn{h} is n the number of observations predicted.
}
\description{
Estimate Observations of \eqn{X} from a Given \code{Y}
}
\details{
this function does inverse prediction, finding observations of \eqn{X} given observations of \eqn{Y}.
}
