% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{center}
\alias{center}
\title{Center a Vector or Matrix}
\usage{
center(X)
}
\arguments{
\item{X}{Numeric vector or matrix}
}
\value{
Centered vector or matrix
}
\description{
Center a Vector or Matrix
}
\details{
This function subtracts the mean from each element of a vector or matrix.
}
\examples{
center(1:10)
}
