% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{confint_EY_h}
\alias{confint_EY_h}
\title{Compute Confidence Intervals for Regression Estimates}
\usage{
confint_EY_h(
  X_h,
  betas,
  MSE,
  covariance,
  n,
  p = NCOL(X_h),
  m = 1,
  alpha = 0.05,
  new = FALSE,
  simultaneous = FALSE
)
}
\arguments{
\item{X_h}{\eqn{h \times p} matrix, or single \eqn{h} length vector, containing data for observations. Be sure to include the
intercept column if present.}

\item{betas}{Vector of estimated coefficients.}

\item{MSE}{Estimated mean squared error.}

\item{covariance}{Covariance matrix of the estimated coefficients.}

\item{n}{Number of observations in the design matrix. Computed if omitted.}

\item{p}{Number of parameters in the model. Computed if omitted.}

\item{m}{Number of new observations to predict.}

\item{alpha}{Alpha level for constructing confidence intervals.}

\item{new}{Logical. Are the values new observations, or were they used to fit the model?
Confidence intervals are narrower in the latter case. Default \code{FALSE}.}

\item{simultaneous}{Logical. Should simultaneous Bonferroni-corrected intervals be constructed instead of separate intervals? Default \code{FALSE}.}
}
\value{
\eqn{h \times 3} matrix of predictions with interval bounds.
}
\description{
Compute Confidence Intervals for Regression Estimates
}
\details{
This function finds the predicted values for given observations predicted
by an estimated linear model. It also computes confidence intervals, the width
varying by whether the observations are new and/or whether simultaneous intervals are desired.
}
