% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{correlate_transform}
\alias{correlate_transform}
\title{Conduct the Correlation Transformation}
\usage{
correlate_transform(X)
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}
}
\value{
The transformed data matrix.
}
\description{
Conduct the Correlation Transformation
}
\details{
Applies the correlation transformation to \code{X} ,
standardizing each variable.
}
