% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_colnames}
\alias{default_colnames}
\title{Assign Generic Column Names to an Object}
\usage{
default_colnames(X, name_base = "X")
}
\arguments{
\item{X}{An object that can receive a \code{colnames} attribute}

\item{name_base}{Stem for each default column name. Default \code{"X"}.}
}
\value{
\code{X}, with its colnames attribute modified as described.
}
\description{
Assign Generic Column Names to an Object
}
\details{
Old column names are retained unless they are the empty string. Any such column names are replaced by new default names numbered from 1. All columns are treated this way if \code{colnames(X)} is \code{NULL}.
}
\examples{
X <- matrix(1:16, nrow = 2)
default_colnames(X)
}
