% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{do_problems}
\alias{do_problems}
\title{Solve Problems Given a Linear Model}
\usage{
do_problems(instance, ..., more_subs = list(), .print = TRUE)
}
\arguments{
\item{instance}{A \code{\link{LinRegProblem}} instance, which is passed as the \code{envir} argument to \code{\link{substitute_call}}.}

\item{...}{Expressions or symbols to evaluate after substitution in the context of \code{instance}.}

\item{more_subs}{Passed directly to \code{\link{substitute_call}}.}

\item{.print}{Logical. Should the results be printed (useful if the output goes in an Rmarkdown file), or returned as a list?}
}
\value{
\code{NULL} if \code{.do_print} is \code{TRUE}, otherwise a list containing results of the expressions.
}
\description{
Solve Problems Given a Linear Model
}
\details{
This function takes arbitrary expressions via \code{...} and evaluates them in the context of a \code{\link{LinRegProblem}} instance. The results are simply printed, if \code{print} is \code{TRUE}, and otherwise returned as a list.
Because \code{\link{substitute_call}} will complete partial calls, this allows for compact specification of solutions.
}
\examples{
X <- LinRegProblem(mtcars[, -1], mtcars[, 1], Y_transform = scale)
do_problems(X, regression_relation())
}
