% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{expected_normal}
\alias{expected_normal}
\title{Find the Expected Values of Residuals Assuming Normality}
\usage{
expected_normal(ranks, mse)
}
\arguments{
\item{ranks}{Ranks of model residuals.}

\item{mse}{Estimated mean squared error of the model.}
}
\value{
Vector of the expected value approximation described above.
}
\description{
Find the Expected Values of Residuals Assuming Normality
}
\details{
This function computes the expected values of model residuals under a normal distribution, as the ordinary model assumptions say they should. The results can be used in normality tests. See \emph{ALSM}, p. 110-111.
}
