% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{extra_SS}
\alias{extra_SS}
\title{Conduct the Generalized \eqn{F} Test}
\usage{
extra_SS(
  X,
  Y,
  reduced_betas,
  full_betas,
  n = NROW(Y),
  p = NCOL(X),
  intercept_fitted = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}

\item{Y}{An \eqn{n}-length vector or \eqn{n \times 1} matrix representing the response.}

\item{reduced_betas}{Vector of indices of beta coefficients for the reduced model, indexed from zero for \eqn{\beta_0}, 1 for \eqn{\beta_1}, etc.}

\item{full_betas}{Vector of indices of beta coefficients for the full model, indexed the same way.}

\item{n}{Number of rows in the design matrix.}

\item{p}{Number of columns in the design matrix.}

\item{intercept_fitted}{Logical. Was an intercept included in the model? Default \code{TRUE}.}

\item{alpha}{Alpha level of the significance test.}
}
\value{
\eqn{F} statistic of the test.
}
\description{
Conduct the Generalized \eqn{F} Test
}
\details{
This function conducts the generalized \eqn{F} test comparing a model with
\eqn{q} parameters to one with \eqn{p} parameters. The extra degrees of freedom are therefore \eqn{q-p}, where \eqn{q -p \geq 1}.
}
