% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_args}
\alias{get_default_args}
\title{Get a Function's Default Arguments}
\usage{
get_default_args(fun, invert = FALSE, ignore_dots = TRUE)
}
\arguments{
\item{fun}{Function object, or the name of a function.}

\item{invert}{Logical. Obtain only non-default arguments?}

\item{ignore_dots}{Logical. Consider \code{...} (if present) a non-default argument?}
}
\value{
Character vector of the names of the specified arguments, \code{NULL} if none meet the conditions.
}
\description{
Get a Function's Default Arguments
}
\details{
This convenience function extracts the default arguments and values from a function, or the non-default arguments if \code{invert = TRUE}.
}
\examples{
get_default_args(lm)

}
