% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{lack_of_fit}
\alias{lack_of_fit}
\title{Conduct the \eqn{F} Test for Lack of Fit}
\usage{
lack_of_fit(
  X,
  Y,
  Y_hat,
  resid = sum_square(Y, Y_hat),
  n = NROW(Y),
  p = NCOL(X),
  alpha = 0.05
)
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}

\item{Y}{An \eqn{n}-length vector or \eqn{n \times 1} matrix representing the response.}

\item{Y_hat}{Vector of fitted values}

\item{resid}{Vector of residuals}

\item{n}{Number of rows in the design matrix. Computed if not provided.}

\item{p}{Number of columns in the design matrix. Computed if not provided.}

\item{alpha}{Alpha level for the significance test.}
}
\value{
Computed \eqn{F} statistic for the test.
}
\description{
Conduct the \eqn{F} Test for Lack of Fit
}
\details{
This function carries out the \eqn{F} test for lack of fit, which
essentially compares the model estimates to the local means of the response
at each distinct observation. The test requires duplicate observations.
}
