% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mse}
\alias{mse}
\title{Compute the Mean Squared Error}
\usage{
mse(Y, Y_hat, n = NROW(Y), p)
}
\arguments{
\item{Y}{Response vector.}

\item{Y_hat}{Vector of predictions.}

\item{n}{Number of rows in design matrix.}

\item{p}{Number of columns in design matrix.}
}
\value{
Estimated mean squared error.
}
\description{
Compute the Mean Squared Error
}
\details{
This function computes the mean squared error, given coefficient and prediction vectors and the parameters for computing degrees of freedom.
}
