% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegProblem.R
\name{new_LinRegProblem}
\alias{new_LinRegProblem}
\title{Construct a LinRegProblem Instance}
\usage{
new_LinRegProblem(
  X,
  Y,
  add_intercept = TRUE,
  X_transform = identity,
  Y_transform = identity
)
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}

\item{Y}{An \eqn{n}-length vector or \eqn{n \times 1} matrix representing the response.}

\item{add_intercept}{Logical. Should an intercept column (a column of 1s) be added to \code{X}?}

\item{X_transform}{One-argument transformation function to apply to \code{X} before fitting the model.
Defaults to \code{\link{identity}}.}

\item{Y_transform}{One-argument transformation function to apply to \code{Y} before fitting the model. Defaults to \code{\link{identity}}.}
}
\value{
A \code{LinRegProblem} object.
}
\description{
Construct a LinRegProblem Instance
}
\details{
This is the constructor function for the \code{LinRegProblem} class. Users
should not use this function and should instead call \code{\link{LinRegProblem}}.
}
