% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{p_value}
\alias{p_value}
\title{Compute Two-sided \eqn{p}-Values under the Normal Distribution}
\usage{
p_value(x, deg_free)
}
\arguments{
\item{x}{A numeric vector}

\item{deg_free}{Degrees of freedom under the null hypothesis distribution}
}
\value{
Vector of computed \eqn{p}-values
}
\description{
Compute Two-sided \eqn{p}-Values under the Normal Distribution
}
\details{
This convenience function computes the probabilities of a vector of values under the distribution assumed by the null hypothesis. I am legally obliged to note this is not remotely the same thing as the probability of a genuine effect.
}
