% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{pseudoinverse}
\alias{pseudoinverse}
\title{Find the Pseudoinverse of a Matrix}
\usage{
pseudoinverse(X)
}
\arguments{
\item{X}{A numeric matrix.}
}
\value{
The pseudoinverse of \eqn{X}.
}
\description{
Find the Pseudoinverse of a Matrix
}
\details{
The pseudoinverse of a matrix is \eqn{V \Sigma^+ U^T} - essentially, its inverse over its row space. It always exists even if no left or right inverse does.
This function is partly based on the \code{\link[MASS]{ginv}}, which does the same thing.
}
\examples{
pseudoinverse(matrix(rnorm(12), ncol = 3))
}
