% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{regression_relation}
\alias{regression_relation}
\title{Conduct the \eqn{F} Test for Existence of a Regression Relation}
\usage{
regression_relation(MSR, MSE, n, p, alpha = 0.05)
}
\arguments{
\item{MSR}{Estimated mean square of regression for the model}

\item{MSE}{Estimated mean squared error for the model}

\item{n}{Number of rows of the data matrix}

\item{p}{Number of parameters in the model}

\item{alpha}{Alpha level to use for the test.}
}
\value{
Value of the computed \eqn{F} statistic
}
\description{
Conduct the \eqn{F} Test for Existence of a Regression Relation
}
\details{
This function tests the null hypothesis that no
regression relation exists, given the relevant mean squares and degrees of freedom.
}
