% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selfname_list}
\alias{selfname_list}
\title{Name a List's Elements After Itself}
\usage{
selfname_list(...)
}
\arguments{
\item{...}{Elements of the list. If unnamed, they are named after themselves}
}
\value{
A list containing the elements of \code{...}, named as described above.
}
\description{
Name a List's Elements After Itself
}
\details{
This function gathers a collection of objects in a list. Names are quoted directly from the expressions passed via \code{....}.
}
\examples{
selfname_list(mtcars, iris, y = 1:5)
}
