% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{simple_fit}
\alias{simple_fit}
\title{Fit a Linear Model}
\usage{
simple_fit(X, betas)
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}

\item{betas}{Vector of coefficients for the model.}
}
\value{
Vector of least-squares estimates.
}
\description{
Fit a Linear Model
}
\details{
This function finds the least-squares estimates given a design matrix and bet coefficients.
}
