% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{simple_ols}
\alias{simple_ols}
\title{Conduct Simple Linear Regression}
\usage{
simple_ols(X, Y, add_intercept = FALSE)
}
\arguments{
\item{X}{An \eqn{n \times m} design matrix, optionally with an intercept column present.}

\item{Y}{An \eqn{n}-length vector or \eqn{n \times 1} matrix representing the response.}

\item{add_intercept}{Logical. Should an intercept column (a column of 1s) be added to \code{X}?}
}
\value{
Vector of estimated beta coefficients.
}
\description{
Conduct Simple Linear Regression
}
\details{
This function finds the minimum-norm least squares solution of \eqn{X^T X \beta = X^T Y}.
}
