% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{SSRDecomp}
\alias{SSRDecomp}
\title{Decompose Sums of Squares within a Linear Model}
\usage{
SSRDecomp(X, Y, reduced_betas, full_betas, intercept_fitted = TRUE)
}
\arguments{
\item{X}{A design matrix for a linear model}

\item{Y}{A response vector for a linear model}

\item{reduced_betas}{Indices for beta coefficients, indexed by 0 (e.g., 0 for \eqn{\beta_0}, etc. ).}

\item{full_betas}{Indices for coefficients of the full model. Defaults to all coefficients.}

\item{intercept_fitted}{Logical. Does the model include an intercept? Default \code{TRUE}.}
}
\value{
Object of \code{SSRDecomp} class. It contains the \eqn{SSR} and \eqn{SSE} of the full and reduced models, along with the coefficient of partial correlation.
}
\description{
Decompose Sums of Squares within a Linear Model
}
\details{
This function decomposes the sums of squares within a reduced and full model and computes the full model's coefficient of partial determination: the portion of \eqn{SSE} of the reduced model it explains.
}
