% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ssto}
\alias{ssto}
\title{Compute the Total Sum of Squares}
\usage{
ssto(X)
}
\arguments{
\item{X}{Vector or matrix.}
}
\value{
Scalar representing the \eqn{SSTO}.
}
\description{
Compute the Total Sum of Squares
}
\details{
Computes the sum of the squared distances of a vector or matrix's elements from its mean.
}
\examples{
with(iris, ssto(cbind(Sepal.Width, Sepal.Length)))
}
