% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{sum_square}
\alias{sum_square}
\title{Compute the Sum of Squares}
\usage{
sum_square(x, y = mean(x))
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric scalar vector of the same length as \code{x}. Defaults to
the mean of \code{x}.}
}
\value{
Sum of squares of the deviation vector.
}
\description{
Compute the Sum of Squares
}
\details{
This function computes the sum of squared differences between \code{x} and \code{y}
}
