% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.LinRegProblem}
\alias{update.LinRegProblem}
\title{Update a \code{LinRegProblem} Object with a New Formula}
\usage{
\method{update}{LinRegProblem}(object, new_formula, ...)
}
\arguments{
\item{object}{A \code{LinRegProblem} object.}

\item{new_formula}{A new model formula. \code{\link{model.matrix}} will attempt to evaluate it in the context of the existing object's column names.}

\item{...}{Additional arguments to \code{\link{LinRegProblem}} to use when creating the new object, or to pass on to methods.}
}
\value{
If successful, the updated \code{LinRegProblem} object.
}
\description{
Update a \code{LinRegProblem} Object with a New Formula
}
\details{
This function refits an existing \code{LinRegProblem} object to a new formula.The formula must be valid in the context of existing column names.
}
\examples{
X <- LinRegProblem(mtcars[, -1], mtcars[, 1], Y_transform = scale)
update(X, Y ~ X1 + X2)
}
