% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Working_Hoteling}
\alias{Working_Hoteling}
\title{Compute Working-Hoteling Confidence Bands}
\usage{
Working_Hoteling(
  X,
  X_h,
  betas,
  Y_h = simple_fit(X_h, betas),
  covariance,
  alpha = 0.05,
  n,
  p,
  MSE
)
}
\arguments{
\item{X}{Design matrix of the linear model.}

\item{X_h}{Matrix of observations to fit the interval for.}

\item{betas}{Estimated regression coefficients.}

\item{Y_h}{\eqn{1 \times h}}

\item{covariance}{\eqn{p \times p} covariance matrix of beta coefficients.}

\item{alpha}{Significance level for the interval.}

\item{n}{Number of rows in the data matrix.}

\item{p}{Number of parameters in the model.}

\item{MSE}{Estimated mean squared error.}
}
\value{
\eqn{h \times 3} matrix of interval bounds
}
\description{
Compute Working-Hoteling Confidence Bands
}
\details{
Given fitted values or enough data to compute them, computes the boundaries of the Working-Hoteling interval estimates.
}
